optim: 22% faster
